refactor: rename project from RoomSystem to OccuPi #87
