+++ backend/src/main/java/com/occupi/app/CorsConfig.java
@@ -1,0 +1,34 @@
+package com.occupi.app;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+/**
+ * Global CORS configuration for HTTP endpoints.
+ *
+ * Allows the frontend (Vite dev server and production builds) to access
+ * backend REST APIs. WebSocket CORS is configured separately in WebSocketConfig.
+ */
+@Configuration
+public class CorsConfig {
+
+    @Bean
+    public WebMvcConfigurer corsConfigurer() {
+        return new WebMvcConfigurer() {
+            @Override
+            public void addCorsMappings(CorsRegistry registry) {
+                registry.addMapping("/**")
+                        .allowedOrigins(
+                                "http://localhost:5173",  // Vite dev server (default)
+                                "http://localhost:3000", // alternative dev
+                                "http://localhost:8181" // InfluxDB
+                        )
+                        .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
+                        .allowedHeaders("*")
+                        .allowCredentials(true);
+            }
+        };
+    }
+}
