+++ backend/src/main/java/com/occupi/feature/receiver/dto/SensorData.java
@@ -1,0 +1,13 @@
+package com.occupi.feature.receiver.dto;
+
+import java.time.Instant;
+
+public record SensorData(
+        String roomId,
+        String sensorId,
+        int count,
+        double confidence,
+        Instant timestamp
+) {}
+
+// Port InfluxDB: 8181
