--- backend/src/main/java/com/roomsystem/app/AppApplication.java
@@ -1,15 +1,0 @@
-package com.roomsystem.app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-@SpringBootApplication
-@ComponentScan(basePackages = "com.roomsystem")
-public class AppApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AppApplication.class, args);
-	}
-
-}
