--- backend/src/main/java/com/roomsystem/feature/authentication/SecurityConfig.java
@@ -1,25 +1,0 @@
-package com.roomsystem.feature.authentication;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Profile("!test")
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-                .csrf(csrf -> csrf.disable())
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/ws/**", "/ws/occupancy/**").permitAll()
-                        .anyRequest().authenticated()
-                );
-        return http.build();
-    }
-}
