--- backend/src/main/java/com/roomsystem/feature/receiver/SensorDataService.java
@@ -1,7 +1,0 @@
-package com.roomsystem.feature.receiver;
-
-import com.roomsystem.feature.receiver.dto.SensorData;
-
-public interface SensorDataService {
-    void process(SensorData data);
-}
