--- backend/src/main/java/com/roomsystem/feature/receiver/dto/SensorData.java
@@ -1,13 +1,0 @@
-package com.roomsystem.feature.receiver.dto;
-
-import java.time.Instant;
-
-public record SensorData(
-        String roomId,
-        String sensorId,
-        int count,
-        double confidence,
-        Instant timestamp
-) {}
-
-// Port InfluxDB: 8181
