+++ backend/src/main/resources/application.yaml
@@ -3,9 +3,13 @@ server:
 
 spring:
   application:
-    name: roomsystem
+    name: occupi
 
 influxdb:
   url: http://localhost:8181
-  database: roomsystem
+  database: occupi
   token: ""
+
+logging:
+  level:
+    com.occupi: DEBUG
