+++ backend/src/test/java/com/occupi/app/AppApplicationTests.java
@@ -1,0 +1,15 @@
+package com.occupi.app;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+@SpringBootTest
+@ActiveProfiles("test")
+class AppApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
