--- backend/src/test/java/com/roomsystem/app/AppApplicationTests.java
@@ -1,15 +1,0 @@
-package com.roomsystem.app;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@SpringBootTest
-@ActiveProfiles("test")
-class AppApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
